refactor(audio): split fabric lease plane and harden test fidelity
Move main-loop lease state and lifecycle into audio_fabric_lease.c
behind a private internal seam; keep the compositor/ISR core focused in
audio_fabric.c. Split parity/compositor and lease scenarios into two
host binaries with a shared harness. Record and assert exact cache
invalidate/flush ordering (mutation-tested), make the source-contract
scan CWD-independent through --srcdir, and add both fabric translation
units to the real firmware C_SRCS manifest (previously missing, which
would have broken the ARM link). Golden parity remains bit-identical;
clean audio and media suites pass.
