+++ test/audio/Makefile
@@ -17,13 +17,19 @@ CONFIG_TARGET := $(BUILD_DIR)/audio_config_test
 MP3_PROBE_TARGET := $(BUILD_DIR)/mp3_frame_probe_test
 MPEG_LAYER2_TARGET := $(BUILD_DIR)/mpeg_layer2_decode_test
 FABRIC_TARGET := $(BUILD_DIR)/audio_fabric_test
+LEASE_TARGET := $(BUILD_DIR)/audio_fabric_lease_test
 TARGETS := $(CONVERT_TARGET) $(CAPTURE_TARGET) $(PROFILE_TARGET) \
 	$(PLAYBACK_TARGET) $(RATE_TARGET) $(DRAIN_TARGET) $(SCENE_TARGET) \
 	$(METER_TARGET) $(CONTROL_TARGET) $(CONFIG_TARGET) \
-	$(MP3_PROBE_TARGET) $(MPEG_LAYER2_TARGET) $(FABRIC_TARGET)
+	$(MP3_PROBE_TARGET) $(MPEG_LAYER2_TARGET) $(FABRIC_TARGET) \
+	$(LEASE_TARGET)
 
 # mock/ shadows ff.h and xil_cache.h for the TUs (zz_config.c) whose
-# real firmware dependencies are FatFs and the Zynq cache API.
+# real firmware dependencies are FatFs and the Zynq cache API. The
+# fabric lease test additionally compiles xil_cache.h in recording
+# mode (-DXIL_CACHE_MOCK_RECORD) so its cache-fidelity scenario can
+# assert the exact invalidate/flush ranges; every other target keeps
+# the no-op stubs.
 CPPFLAGS += -I$(MOCK_DIR) -I$(SRC_DIR)
 
 .PHONY: test clean check-tables
@@ -40,13 +46,17 @@ test: $(TARGETS)
 	$(METER_TARGET)
 	$(CONTROL_TARGET)
 	$(MPEG_LAYER2_TARGET)
-	$(FABRIC_TARGET)
+	$(FABRIC_TARGET) --srcdir $(SRC_DIR)
+	$(LEASE_TARGET)
 
 # U2 fabric compositor: the characterization golden (pump_golden.c, a
 # verbatim capture of the pre-fabric pump) drives the parity scenarios
-# against audio_fabric.c through the host seam stubs in the test TU.
+# against audio_fabric.c + audio_fabric_lease.c through the host seam
+# stubs in the shared harness (fabric_test_common.c).
 $(FABRIC_TARGET): audio_fabric_test.c pump_golden.c pump_golden.h \
-	$(SRC_DIR)/audio_fabric.c $(SRC_DIR)/audio_fabric.h \
+	fabric_test_common.c fabric_test_common.h \
+	$(SRC_DIR)/audio_fabric.c $(SRC_DIR)/audio_fabric_lease.c \
+	$(SRC_DIR)/audio_fabric.h $(SRC_DIR)/audio_fabric_internal.h \
 	$(SRC_DIR)/audio_convert.c $(SRC_DIR)/audio_convert.h \
 	$(SRC_DIR)/audio_convert_tables.c $(SRC_DIR)/audio_convert_tables.h \
 	$(SRC_DIR)/audio_playback_frontier.h $(SRC_DIR)/ax.h \
@@ -55,7 +65,27 @@ $(FABRIC_TARGET): audio_fabric_test.c pump_golden.c pump_golden.h \
 	@mkdir -p $(BUILD_DIR)
 	$(CC) $(CPPFLAGS) $(CFLAGS) -DAUDIO_FABRIC_HOST_TEST \
 		-o $@ audio_fabric_test.c \
-		pump_golden.c $(SRC_DIR)/audio_fabric.c \
+		pump_golden.c fabric_test_common.c \
+		$(SRC_DIR)/audio_fabric.c $(SRC_DIR)/audio_fabric_lease.c \
+		$(SRC_DIR)/audio_convert.c $(SRC_DIR)/audio_convert_tables.c -lm
+
+# U3 mailbox lease plane: the SDK-facing lease API against the same two
+# firmware TUs, with xil_cache.h compiled in recording mode so the
+# cache-fidelity scenario asserts the exact invalidate/flush ranges.
+$(LEASE_TARGET): audio_fabric_lease_test.c \
+	fabric_test_common.c fabric_test_common.h \
+	$(SRC_DIR)/audio_fabric.c $(SRC_DIR)/audio_fabric_lease.c \
+	$(SRC_DIR)/audio_fabric.h $(SRC_DIR)/audio_fabric_internal.h \
+	$(SRC_DIR)/audio_convert.c $(SRC_DIR)/audio_convert.h \
+	$(SRC_DIR)/audio_convert_tables.c $(SRC_DIR)/audio_convert_tables.h \
+	$(SRC_DIR)/audio_playback_frontier.h $(SRC_DIR)/ax.h \
+	$(SRC_DIR)/memorymap.h $(SRC_DIR)/sdk_mailbox.h \
+	$(MOCK_DIR)/xil_cache.h
+	@mkdir -p $(BUILD_DIR)
+	$(CC) $(CPPFLAGS) $(CFLAGS) -DAUDIO_FABRIC_HOST_TEST \
+		-DXIL_CACHE_MOCK_RECORD \
+		-o $@ audio_fabric_lease_test.c fabric_test_common.c \
+		$(SRC_DIR)/audio_fabric.c $(SRC_DIR)/audio_fabric_lease.c \
 		$(SRC_DIR)/audio_convert.c $(SRC_DIR)/audio_convert_tables.c -lm
 
 $(CONVERT_TARGET): audio_convert_test.c $(SRC_DIR)/audio_convert.c \
